question1/step1__stepfunction__scrape_structure/s2__aws_batch__small_batch_html_parsing/requirements.txt
aioboto3==9.4.0
aiobotocore==2.1.2
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
boto3==1.20.24
botocore==1.23.24
bs4==0.0.1
charset-normalizer==2.0.12
frozenlist==1.3.0
html5lib==1.1
idna==3.3
jmespath==0.10.0
lxml==4.8.0
multidict==6.0.2
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.0
yarl==1.7.2
